+++ priceservice/src/main/java/com/example/priceservice/dto/PriceDTO.java
@@ -1,0 +1,13 @@
+package com.example.priceservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PriceDTO {
+    private Double amount;
+    private String currency;
+}
