+++ priceservice/src/main/java/com/example/priceservice/repository/PriceRepository.java
@@ -1,0 +1,12 @@
+package com.example.priceservice.repository;
+
+import com.example.priceservice.model.Price;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface PriceRepository extends JpaRepository<Price, Long> {
+    Optional<Price> findByProductId(Long productId);
+}
