--- zip (1)/src/main/java/com/example/priceservice/model/Price.java
@@ -1,19 +1,0 @@
-package com.example.priceservice.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Price {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Long productId;
-    private Double amount;
-    private String currency;
-}
