--- zip (1)/src/main/java/com/example/priceservice/service/PriceService.java
@@ -1,24 +1,0 @@
-package com.example.priceservice.service;
-
-import com.example.priceservice.model.Price;
-import com.example.priceservice.repository.PriceRepository;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.stereotype.Service;
-
-
-@Service
-@RequiredArgsConstructor
-public class PriceService {
-    private final PriceRepository priceRepository;
-
-    public Price getPriceByProductId(Long productId) {
-        return priceRepository.findByProductId(productId)
-                .orElseThrow(() -> new RuntimeException("Precio no encontrado"));
-    }
-
-    public Price assignPrice(Price price) {
-        return priceRepository.save(price);
-    }
-}
