--- zip/src/main/java/com/example/productservice/Application.java
@@ -1,27 +1,0 @@
-package com.example.productservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Bean;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-@SpringBootApplication
-@EnableFeignClients
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-
-    @Bean
-    OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info().title("Product Service API")
-                        .version("1.0")
-                        .description("API para la gestión de productos"));
-    }
-
-}
