--- zip/src/main/java/com/example/productservice/repository/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.example.productservice.repository;
-
-import com.example.productservice.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
