--- zip/src/main/resources/application.properties
@@ -1,34 +1,0 @@
-spring.security.user.name=admin
-spring.security.user.password=admin123
-
-# Configuración de JWT
-security.jwt.secret=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJUZXN0RUdTTCIsImlhdCI6MTc0MTk0NDI4OSwiZXhwIjoxNzczNDgwMjg4LCJhdWQiOiJsb2NhbGhvc3QiLCJzdWIiOiJlcnNhbG8xMjNAZ21haWwiLCJ1c2VybmFtZSI6Im15S2V5QWxpYXMiLCJwYXNzd29yZCI6Im15cGFzc3dvcmQifQ.AruhpOMyW7vJljUw_ix7KUXskGSngT29CRn9kAEzXlI
-security.jwt.expiration=86400000
-
-spring.datasource.driverClassName=org.h2.Driver
-spring.h2.console.path=/api/h2-console
-spring.datasource.url=jdbc:h2:mem:productdb
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=update
-spring.h2.console.enabled=true
-
-server.port=8443
-
-spring.application.name=product-service
-
-spring.devtools.livereload.enabled=true
-
-feign.client.config.default.connectTimeout=5000
-feign.client.config.default.readTimeout=5000
-
-server.ssl.key-store=classpath:keystore.p12
-server.ssl.key-store-password=mypassword
-server.ssl.key-store-type=PKCS12
-server.ssl.key-alias=myKeyAlias
-server.ssl.key-password=mypassword
-server.ssl.enabled=true
-server.servlet.context-path=/api/products
-feign.client.config.price-service.url=https://localhost:8081/api/prices
